+++ src/main/java/net/frozenblock/lib/cape/client/impl/ClientCapeData.java
@@ -27,10 +27,10 @@ import net.fabricmc.fabric.api.client.event.lifecycle.v1.ClientEntityEvents;
 import net.fabricmc.fabric.api.client.event.lifecycle.v1.ClientLifecycleEvents;
 import net.fabricmc.fabric.api.client.networking.v1.ClientPlayConnectionEvents;
 import net.fabricmc.fabric.api.client.networking.v1.ClientPlayNetworking;
+import net.frozenblock.lib.cape.api.CapeUtil;
 import net.frozenblock.lib.cape.impl.Cape;
 import net.frozenblock.lib.cape.impl.networking.CapeCustomizePacket;
 import net.frozenblock.lib.config.frozenlib_config.FrozenLibConfig;
-import net.frozenblock.lib.registry.api.FrozenRegistry;
 import net.minecraft.client.Minecraft;
 import net.minecraft.client.multiplayer.ClientLevel;
 import net.minecraft.resources.ResourceLocation;
@@ -38,28 +38,22 @@ import org.jetbrains.annotations.NotNull;
 
 @Environment(EnvType.CLIENT)
 public class ClientCapeData {
-	private static final Map<UUID, Cape> CAPES_IN_SERVER = new HashMap<>();
+	private static final Map<UUID, Cape> CAPES_IN_WORLD = new HashMap<>();
 
 	public static Optional<ResourceLocation> getCapeTexture(UUID uuid) {
-		return Optional.ofNullable(CAPES_IN_SERVER.get(uuid)).map(Cape::texture);
+		return Optional.ofNullable(CAPES_IN_WORLD.get(uuid)).map(Cape::texture);
 	}
 
 	public static void setCapeForUUID(UUID uuid, ResourceLocation capeId) {
-		Cape cape = FrozenRegistry.CAPE.getValue(capeId);
-		if (cape == null) {
-			removeCapeForUUID(uuid);
-		} else {
-			CAPES_IN_SERVER.put(uuid, cape);
-			setPlayerCapeTexture(uuid, Optional.of(cape));
-		}
+		CapeUtil.getCape(capeId).ifPresentOrElse(cape -> setPlayerCape(uuid, Optional.of(cape)), () -> removeCapeForUUID(uuid));
 	}
 
 	public static void removeCapeForUUID(UUID uuid) {
-		CAPES_IN_SERVER.remove(uuid);
-		setPlayerCapeTexture(uuid, Optional.empty());
+		setPlayerCape(uuid, Optional.empty());
 	}
 
-	private static void setPlayerCapeTexture(UUID uuid, @NotNull Optional<Cape> cape) {
+	private static void setPlayerCape(UUID uuid, @NotNull Optional<Cape> cape) {
+		cape.ifPresentOrElse(cape1 -> CAPES_IN_WORLD.put(uuid, cape1), () -> CAPES_IN_WORLD.remove(uuid));
 		ClientLevel level = Minecraft.getInstance().level;
 		if (level != null && level.getPlayerByUUID(uuid) instanceof PlayerCapeInterface capeInterface) {
 			capeInterface.frozenLib$setCape(cape.map(Cape::texture).orElse(null));
@@ -67,13 +61,10 @@ public class ClientCapeData {
 	}
 
 	public static void init() {
-		ClientLifecycleEvents.CLIENT_STOPPING.register(client -> {
-			CAPES_IN_SERVER.clear();
-		});
-		ClientPlayConnectionEvents.DISCONNECT.register((clientPacketListener, minecraft) -> CAPES_IN_SERVER.clear());
-		ClientPlayConnectionEvents.JOIN.register((clientPacketListener, packetSender, minecraft) -> {
-			ClientPlayNetworking.send(CapeCustomizePacket.createPacket(minecraft.getUser().getProfileId(), ResourceLocation.parse(FrozenLibConfig.get().cape)));
-		});
+		ClientLifecycleEvents.CLIENT_STOPPING.register(client -> CAPES_IN_WORLD.clear());
+		ClientPlayConnectionEvents.DISCONNECT.register((clientPacketListener, minecraft) -> CAPES_IN_WORLD.clear());
+		ClientPlayConnectionEvents.JOIN.register((clientPacketListener, packetSender, minecraft) ->
+			ClientPlayNetworking.send(CapeCustomizePacket.createPacket(minecraft.getUser().getProfileId(), ResourceLocation.parse(FrozenLibConfig.get().cape))));
 		ClientEntityEvents.ENTITY_LOAD.register((entity, clientLevel) -> {
 			if (entity instanceof PlayerCapeInterface capeInterface) {
 				getCapeTexture(entity.getUUID()).ifPresent(capeInterface::frozenLib$setCape);